add: final upload

+++ app/src/main/java/com/school/codes/ric/mobileappsproject/ui/AddCourseFragment.java
@@ -1,13 +1,16 @@
 package com.school.codes.ric.mobileappsproject.ui;
 
+import android.app.AlarmManager;
 import android.app.DatePickerDialog;
+import android.app.PendingIntent;
 import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.NotificationCompat;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -35,6 +38,13 @@ import java.util.Calendar;
 import java.util.List;
 import java.util.Objects;
 
+import static android.content.Context.ALARM_SERVICE;
+import static com.school.codes.ric.mobileappsproject.util.Constants.NOTIFICATION;
+import static com.school.codes.ric.mobileappsproject.util.Constants.NOTIFICATION_ID;
+import static com.school.codes.ric.mobileappsproject.util.Constants.NOTIFICATION_ID_OFFSET;
+import static com.school.codes.ric.mobileappsproject.util.Constants.NOTIFICATION_TITLE_END;
+import static com.school.codes.ric.mobileappsproject.util.Constants.NOTIFICATION_TITLE_START;
+
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
@@ -46,9 +56,11 @@ import java.util.Objects;
 public class AddCourseFragment extends Fragment {
 
     public static final String TAG = AddCourseFragment.class.getSimpleName();
+    private static final String ID = "ID";
 
     private OnFragmentInteractionListener mListener;
     private CourseDAO courseDAO;
+    private AssessmentDAO assessmentDAO;
 
     private TextView startDate;
     private TextView endDate;
@@ -62,7 +74,7 @@ public class AddCourseFragment extends Fragment {
     private EditText mentorEditText;
     private CourseRO course;
     private EditText titleEditText;
-    private AssessmentDAO assessmentDAO;
+    private boolean isNewCourse;
 
     public AddCourseFragment() {
         // Required empty public constructor
@@ -74,8 +86,12 @@ public class AddCourseFragment extends Fragment {
      *
      * @return A new instance of fragment AddTermFragment.
      */
-    public static AddCourseFragment newInstance() {
-        return new AddCourseFragment();
+    public static AddCourseFragment newInstance(int id) {
+        AddCourseFragment fragment = new AddCourseFragment();
+        Bundle args = new Bundle();
+        args.putInt(ID, id);
+        fragment.setArguments(args);
+        return fragment;
     }
 
     @Override
@@ -189,20 +205,24 @@ public class AddCourseFragment extends Fragment {
         cancelButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                List<AssessmentRO> tempAssessments = null;
 
-                try {
-                    tempAssessments = assessmentDAO.getAllAssessmentsForCourse(course.getId());
-                } catch (ParseException e) {
-                    e.printStackTrace();
+                if (isNewCourse) {
+                    List<AssessmentRO> tempAssessments = null;
+
+                    try {
+                        tempAssessments = assessmentDAO.getAllAssessmentsForCourse(course.getId());
+                    } catch (ParseException e) {
+                        e.printStackTrace();
+                    }
+
+                    assert tempAssessments != null;
+                    for (AssessmentRO a : tempAssessments) {
+                        assessmentDAO.disassociate(a);
+                    }
+
+                    courseDAO.delete(courseDAO.getLastId());
                 }
 
-                assert tempAssessments != null;
-                for (AssessmentRO a : tempAssessments) {
-                    assessmentDAO.disassociate(a);
-                }
-
-                courseDAO.delete(courseDAO.getLastId());
                 goToHomePage();
             }
         });
@@ -242,7 +262,17 @@ public class AddCourseFragment extends Fragment {
                     course.setStatus(statusSpinner.getSelectedItem().toString());
 
                     courseDAO.update(course);
-                    Log.d(TAG, "course update=" + course.getId());
+
+                    scheduleNotification(NOTIFICATION_TITLE_START,
+                            course.getTitle(),
+                            course.getId(),
+                            course.getStartAlert());
+
+                    scheduleNotification(NOTIFICATION_TITLE_END,
+                            course.getTitle(),
+                            course.getId() + NOTIFICATION_ID_OFFSET,
+                            course.getEndAlert());
+
                     onCourseSave(course.getId());
 
                     Toast.makeText(getContext(), "Course created successfully", Toast.LENGTH_SHORT)
@@ -255,39 +285,124 @@ public class AddCourseFragment extends Fragment {
 
         try {
 
-            course = new CourseRO("",
-                    DateUtils.convertStringToTimestamp("01-Jan-1990"),
-                    DateUtils.convertStringToTimestamp("01-Jan-1990"),
-                    Constants.PLAN_STATUS,
-                    "",
-                    "",
-                    DateUtils.convertStringToTimestamp("01-Jan-1990"),
-                    DateUtils.convertStringToTimestamp("01-Jan-1990")
-            );
-            courseDAO.add(course);
-            course = courseDAO.get(courseDAO.getLastId());
-            Log.d(TAG, "course added=" + course.getId());
+            if (getArguments().getInt(ID) == 0) {
+                isNewCourse = true;
 
-            RecyclerView associateNewCoursesRecycler =
-                    v.findViewById(R.id.associateNewCoursesRecycler);
-            RecyclerView.LayoutManager manager = new LinearLayoutManager(getContext());
-            assessmentDAO = new AssessmentDAO(getContext());
-            AssociateAssessmentAdapter adapter = new AssociateAssessmentAdapter(assessmentDAO.getAll(),
-                    course,
-                    associateNewCoursesRecycler,
-                    getContext());
+                course = new CourseRO("",
+                        DateUtils.convertStringToTimestamp("01-Jan-1990"),
+                        DateUtils.convertStringToTimestamp("01-Jan-1990"),
+                        Constants.PLAN_STATUS,
+                        "",
+                        "",
+                        DateUtils.convertStringToTimestamp("01-Jan-1990"),
+                        DateUtils.convertStringToTimestamp("01-Jan-1990")
+                );
+                courseDAO.add(course);
+                course = courseDAO.get(courseDAO.getLastId());
+            } else {
+                isNewCourse = false;
 
-            associateNewCoursesRecycler.setLayoutManager(manager);
-            associateNewCoursesRecycler.setAdapter(adapter);
+                course = courseDAO.get((getArguments().getInt(ID)));
+
+                titleEditText.setText(course.getTitle());
+                start = course.getStart();
+                startDate.setText(DateUtils.convertTimestampToString(start));
+                end = course.getEnd();
+                endDate.setText(DateUtils.convertTimestampToString(end));
+                endAlert = course.getEndAlert();
+                endAlertDate.setText(DateUtils.convertTimestampToString(endAlert));
+                startAlert = course.getStartAlert();
+                startAlertDate.setText(DateUtils.convertTimestampToString(startAlert));
+                mentorEditText.setText(course.getMentor());
+                if (course.getNotes() != null) {
+                    notesEditText.setText(course.getNotes());
+                }
+
+                setSpinner(statusSpinner);
+            }
+
+            setUpRecycler(v);
 
         } catch (ParseException e) {
-            Log.e(TAG, "ERROR=" + e.getMessage());
-            e.printStackTrace(); //todo: handle this
+            e.printStackTrace();
             return null;
         }
 
         return v;
 
+    }
+
+    private void scheduleNotification(String title, String content, int id, Timestamp alert) {
+
+
+        // Set up intent to open TaskListFragment
+        Intent mainActivityIntent = new Intent(getContext(), MainActivity.class);
+        PendingIntent mainActivityPendingIntent =
+                PendingIntent.getActivity(getContext(), id, mainActivityIntent,
+                        PendingIntent.FLAG_UPDATE_CURRENT);
+
+        // build notification
+        NotificationCompat.Builder notification =
+                new NotificationCompat.Builder(getContext());
+        notification.setContentTitle(title);
+        notification.setContentText(content);
+        notification.setDefaults(NotificationCompat.DEFAULT_SOUND);
+        notification.setSmallIcon(R.drawable.notify);
+        notification.setAutoCancel(true);
+        notification.setNumber(1);
+        notification.setContentIntent(mainActivityPendingIntent);
+
+
+        // create intent to add task values for notification
+        Intent notificationIntent = new Intent(getContext(),
+                NotifPublisher.class);
+        notificationIntent.putExtra(NOTIFICATION_ID, id);
+        notificationIntent.putExtra(NOTIFICATION, notification.build());
+
+        // call alarm manager to schedule notification
+        PendingIntent pendingIntent = PendingIntent.getBroadcast(getContext()
+                , id, notificationIntent, PendingIntent.FLAG_UPDATE_CURRENT);
+
+        AlarmManager alarmManager =
+                (AlarmManager) getContext().getSystemService(ALARM_SERVICE);
+        assert alarmManager != null;
+        alarmManager.setExact(AlarmManager.RTC_WAKEUP,
+                alert.getTime(),
+                pendingIntent);
+
+    }
+
+    private void setSpinner(Spinner statusSpinner) {
+        String status = course.getStatus();
+        int index;
+        switch (status) {
+            case Constants.IN_PROGRESS_STATUS:
+                index = 1;
+                break;
+            case Constants.COMPLETE_STATUS:
+                index = 2;
+                break;
+            case Constants.DROPPED_STATUS:
+                index = 3;
+                break;
+            default:
+                index = 0;
+        }
+        statusSpinner.setSelection(index);
+    }
+
+    private void setUpRecycler(View v) throws ParseException {
+        RecyclerView associateNewCoursesRecycler =
+                v.findViewById(R.id.associateNewCoursesRecycler);
+        RecyclerView.LayoutManager manager = new LinearLayoutManager(getContext());
+        assessmentDAO = new AssessmentDAO(getContext());
+        AssociateAssessmentAdapter adapter = new AssociateAssessmentAdapter(assessmentDAO.getAll(),
+                course,
+                associateNewCoursesRecycler,
+                getContext());
+
+        associateNewCoursesRecycler.setLayoutManager(manager);
+        associateNewCoursesRecycler.setAdapter(adapter);
     }
 
     public void onCourseSave(int id) {
